Added submodule for FreeRTOS kernel V7.0.2.
The files in Source\ of the kernel in version 7.0.2 are identical to those in src\freertos\source\.
Some files in src\freertos\source\ are not apparent in Source\ of the kernel in version 7.0.2. Those have the identifier for version 8.2.2 in their header.
There is a mix of kernel versions in src\freertos\source\.